+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+# v0.3.1
+
 * Added task `mix wechat_pay.get_sandbox_signkey` to get the Sandbox API Key.
 * Fixes wrong return value of `WechatPay.Helper.get_sandbox_signkey/2`.